Add files via upload
Final (hopefully) upload of the chase_location code. Adjusted teamfile_location to correctly reflect the directory where they will be placed during the actual simulation. Added code to check to see if a selected team name is already in use. Modified the print statement that provides links to the user for placefile addition so that it doesn't need to be edited separately. Added a section with the links for polling configuration, warning configuration, and LSR placefile. Also added welcome text. Modified the necessary variables for final chase simulation values.